requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-cors-middleware==1.3.1
django-extensions==2.1.6
django-heroku==0.3.1
psycopg2==2.7.7
dj-database-url==0.5.0
django-heroku==0.3.1
whitenoise==4.1.2
gunicorn==19.9.0
factory-boy==2.11.1
coverage==4.5.3
graphene_django==2.2.0
requests==2.21.0
PyJWT==1.7.1 